hardware/requirements.txt
opencv-python-headless>=4.10.0.84
python-dotenv>=1.0.1
websockets>=13.0
